Exemplos da aula do dia 23 de Agosto

+++ Aula-T02/Main_03.java
@@ -6,7 +6,7 @@ public class Main_03 {
          * tipo nome = valor Inicial;
          */
         
-        byte bVar = 127; // 1 byte - Stores whole numbers from -128 to 127
+        byte bVar = (byte)127; // 1 byte - Stores whole numbers from -128 to 127
         short sVar = 32_767; // 2 bytes - Stores whole numbers from -32,768 to 32,767
         int iVar = 2_147_483_647; // 4 bytes - Stores whole numbers from -2,147,483,648 to 2,147,483,647
         // 8 bytes - Stores whole numbers from -9,223,372,036,854,775,808 to 9,223,372,036,854,775,807
